+++ components/style/build_gecko.rs
@@ -353,6 +353,7 @@ mod bindings {
             "nsCursorImage",
             "nsFont",
             "nsIAtom",
+            "nsIURI",
             "nsMainThreadPtrHandle",
             "nsMainThreadPtrHolder",
             "nsMargin",
@@ -613,6 +614,7 @@ mod bindings {
             "nsCursorImage",
             "nsFont",
             "nsIAtom",
+            "nsIURI",
             "nsMediaFeature",
             "nsRestyleHint",
             "nsStyleBackground",
