+++ components/style/servo/mod.rs
@@ -9,3 +9,4 @@
 pub mod media_queries;
 pub mod restyle_damage;
 pub mod selector_parser;
+pub mod url;
